uploadersincat: Apply noHTML() when outputting category title, not when accepting it
It is perfectly valid to have a category named `Foo:bar:baz`,
however, this tool would pass it to category class as `Foo&colon;bar&colon;baz`,
which is also perfectly valid, but not what the user intended to input. It doesn't hurt
to pass valid HTML to the database - PDO's escpaing should handle that well and not let
any characters capable to cause SQL injection to reach the database.

To prevent attackers from naming category `<script>alert('foo');</script>`,
let's escape the category name when we output it. At that point, outputing
`&colon;` instead of colons doesn't hurt anything.

Fixes #5.